Refactor API usage and remove unused utility modules
Replaced all imports from '@/lib/api' with '@/lib/api-service' or '@/lib/api-client' across the codebase, consolidating API logic. Updated environment variable checks from NEXT_PUBLIC_DEBUG_AUTH to NEXT_PUBLIC_DEV_MODE. Removed unused utility modules in lib/ (accessibility, design-system, mobile, permissions, validation). Refactored contract transaction CSV export to use the new apiClient.download method. Renamed lib/api.ts to lib/api-service.ts and added a new lib/api-client.ts for API client logic.
