+++ components/features/notifications/notification-dropdown.tsx
@@ -43,7 +43,7 @@ import {
 } from "@/components/ui/alert-dialog";
 import { Badge } from "@/components/ui/badge";
 import { ScrollArea } from "@/components/ui/scroll-area";
-import apiClient from "@/lib/api";
+import apiClient from "@/lib/api-client";
 import { useNotificationContext } from "@/context/notification-context";
 import {
   NotificationItem,
