+++ components/features/roles/role-table.tsx
@@ -31,7 +31,7 @@ import {
   KeyRound,
   Users,
 } from "lucide-react";
-import { deleteRole } from "@/lib/api";
+import { deleteRole } from "@/lib/api-service";
 import {
   showDeleteConfirmation,
   showDeleteSuccess,
@@ -254,55 +254,57 @@ export function RoleTable({
                       ไม่พบข้อมูลตำแหน่ง
                     </TableCell>
                   </TableRow>
-                ) : (filteredRoles.map((role) => (
-                  <TableRow key={role.id} className="hover:bg-slate-50">
-                    <TableCell className="font-medium text-slate-800">
-                      {role.name}
-                    </TableCell>
-                    <TableCell className="text-slate-600 max-w-md truncate">
-                      {role.description}
-                    </TableCell>
-                    <TableCell>
-                      <span className="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium bg-slate-100 text-slate-700">
-                        {role.userCount} คน
-                      </span>
-                    </TableCell>
-                    <TableCell className="text-slate-600 text-sm">
-                      {new Date(role.updatedAt).toLocaleDateString("th-TH")}
-                    </TableCell>
-                    <TableCell>
-                      <div className="flex items-center justify-center space-x-1">
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          onClick={() => {
-                            setSelectedRole(role);
-                            setIsViewDialogOpen(true);
-                          }}
-                          className="text-blue-600 hover:text-blue-800 hover:bg-slate-100"
-                        >
-                          <Eye className="w-4 h-4" />
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          onClick={() => onEditRole(role)}
-                          className="text-slate-600 hover:text-slate-800 hover:bg-slate-100"
-                        >
-                          <Edit className="w-4 h-4" />
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          onClick={() => handleDeleteRole(role.id, role.name)}
-                          className="text-red-600 hover:text-red-800 hover:bg-red-50"
-                        >
-                          <Trash2 className="w-4 h-4" />
-                        </Button>
-                      </div>
-                    </TableCell>
-                  </TableRow>
-                )))}
+                ) : (
+                  filteredRoles.map((role) => (
+                    <TableRow key={role.id} className="hover:bg-slate-50">
+                      <TableCell className="font-medium text-slate-800">
+                        {role.name}
+                      </TableCell>
+                      <TableCell className="text-slate-600 max-w-md truncate">
+                        {role.description}
+                      </TableCell>
+                      <TableCell>
+                        <span className="inline-flex items-center px-2 py-1 rounded-full text-xs font-medium bg-slate-100 text-slate-700">
+                          {role.userCount} คน
+                        </span>
+                      </TableCell>
+                      <TableCell className="text-slate-600 text-sm">
+                        {new Date(role.updatedAt).toLocaleDateString("th-TH")}
+                      </TableCell>
+                      <TableCell>
+                        <div className="flex items-center justify-center space-x-1">
+                          <Button
+                            variant="ghost"
+                            size="sm"
+                            onClick={() => {
+                              setSelectedRole(role);
+                              setIsViewDialogOpen(true);
+                            }}
+                            className="text-blue-600 hover:text-blue-800 hover:bg-slate-100"
+                          >
+                            <Eye className="w-4 h-4" />
+                          </Button>
+                          <Button
+                            variant="ghost"
+                            size="sm"
+                            onClick={() => onEditRole(role)}
+                            className="text-slate-600 hover:text-slate-800 hover:bg-slate-100"
+                          >
+                            <Edit className="w-4 h-4" />
+                          </Button>
+                          <Button
+                            variant="ghost"
+                            size="sm"
+                            onClick={() => handleDeleteRole(role.id, role.name)}
+                            className="text-red-600 hover:text-red-800 hover:bg-red-50"
+                          >
+                            <Trash2 className="w-4 h-4" />
+                          </Button>
+                        </div>
+                      </TableCell>
+                    </TableRow>
+                  ))
+                )}
               </TableBody>
             </Table>
           </div>
