+++ components/widgets/asset-type/asset-type-summary.tsx
@@ -9,7 +9,7 @@ import {
   ChartTooltipContent,
 } from "@/components/ui/chart";
 import { PieChart, Pie, Cell, ResponsiveContainer } from "recharts";
-import apiClient from "@/lib/api";
+import apiClient from "@/lib/api-client";
 import { useWidgetRegistration } from "@/context/widget-context";
 import { useOptimizedMemo, usePerformanceMonitor } from "@/lib/performance";
