+++ components/widgets/dashboard/gold-price.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Coins, Info } from "lucide-react";
-import apiClient from "@/lib/api";
+import apiClient from "@/lib/api-client";
 import { useWidgetRegistration } from "@/context/widget-context";
 import type { GoldPrice } from "@/types/dashboard";
 
@@ -38,7 +38,7 @@ export const GoldPriceCard = () => {
         setError(errorMessage);
 
         // Log error ใน development mode
-        if (process.env.NEXT_PUBLIC_DEBUG_AUTH === "true") {
+        if (process.env.NEXT_PUBLIC_DEV_MODE === "true") {
           console.error("❌ Failed to fetch gold price:", err);
         }
       } finally {
