+++ lib/api-client.ts
@@ -1,0 +1,338 @@
+import { ApiResponse, ApiErrorResponse } from "@/types/api";
+
+// Types สำหรับ fetch API
+interface FetchConfig extends RequestInit {
+  timeout?: number;
+}
+
+interface FetchError extends Error {
+  status?: number;
+  response?: ApiErrorResponse;
+}
+
+/**
+ * Get base URL from environment variable
+ */
+const getBaseUrl = (): string => {
+  return process.env.NEXT_PUBLIC_API_URL || "";
+};
+
+/**
+ * API Client Class - Singleton pattern สำหรับการจัดการ HTTP requests
+ */
+class ApiClient {
+  private readonly baseURL: string;
+  private readonly timeout: number;
+
+  constructor() {
+    this.baseURL = getBaseUrl();
+    this.timeout = 30000; // 30 วินาที timeout
+
+    // ตรวจสอบว่า baseURL ไม่ว่าง
+    if (!this.baseURL) {
+      console.error("❌ API_URL is not set! Check environment variables.");
+    }
+  }
+
+  /**
+   * สร้าง default headers พร้อม configuration สำหรับ httpOnly cookies
+   */
+  private getDefaultHeaders(): Record<string, string> {
+    const headers: Record<string, string> = {
+      "Content-Type": "application/json",
+      Accept: "application/json",
+    };
+
+    // เพิ่ม Authorization header ถ้ามี token ใน localStorage
+    if (typeof window !== "undefined") {
+      const token = localStorage.getItem("accessToken");
+      if (token) {
+        headers.Authorization = `Bearer ${token}`;
+      }
+    }
+
+    return headers;
+  }
+
+  /**
+   * สร้าง default config สำหรับ fetch
+   */
+  private getDefaultConfig(): RequestInit {
+    return {
+      credentials: "include", // ส่ง cookies และ authentication headers
+      mode: "cors", // เพิ่ม CORS mode
+      cache: "no-cache", // ป้องกัน cache issues
+      headers: {
+        ...this.getDefaultHeaders(),
+        // เพิ่ม headers เพื่อรองรับ CORS และ Azure
+        "Access-Control-Allow-Credentials": "true",
+        "X-Requested-With": "XMLHttpRequest",
+      },
+    };
+  }
+
+  /**
+   * จัดการ timeout สำหรับ fetch request
+   */
+  private async fetchWithTimeout(
+    url: string,
+    config: RequestInit & { timeout?: number }
+  ): Promise<Response> {
+    const { timeout = this.timeout, ...fetchConfig } = config;
+
+    const controller = new AbortController();
+    const timeoutId = setTimeout(() => controller.abort(), timeout);
+
+    try {
+      const response = await fetch(url, {
+        ...fetchConfig,
+        signal: controller.signal,
+      });
+
+      clearTimeout(timeoutId);
+      return response;
+    } catch (error) {
+      console.error("❌ fetch error:", {
+        url,
+        error: error instanceof Error ? error.message : error,
+        name: error instanceof Error ? error.name : "Unknown",
+        cause: error instanceof Error ? error.cause : undefined,
+      });
+
+      clearTimeout(timeoutId);
+      throw error;
+    }
+  }
+
+  /**
+   * จัดการ Response Errors อย่างเป็นระบบ
+   */
+  private async handleResponse<T>(response: Response): Promise<ApiResponse<T>> {
+    if (!response.ok) {
+      let errorData: ApiErrorResponse | null = null;
+
+      try {
+        errorData = await response.json();
+      } catch {
+        // ถ้า parse JSON ไม่ได้
+      }
+
+      // Log error ใน development mode
+      if (process.env.NEXT_PUBLIC_DEV_MODE === "true") {
+        console.error("❌ API Error:", {
+          status: response.status,
+          message: errorData?.message,
+          url: response.url,
+        });
+      }
+
+      // จัดการ error ตาม status code
+      switch (response.status) {
+        case 401:
+          console.warn("🔐 Unauthorized: Token may be expired");
+          break;
+        case 403:
+          console.warn("🚫 Forbidden: Insufficient permissions");
+          break;
+        case 404:
+          console.warn("🔍 Not Found: Resource not found");
+          break;
+        case 429:
+          console.warn("⏳ Rate Limited: Too many requests");
+          break;
+        case 500:
+          console.error("🔥 Server Error: Internal server error");
+          break;
+        default:
+          console.error("❓ Unknown Error:", response.status);
+      }
+
+      const error: FetchError = new Error(
+        errorData?.message || `HTTP Error ${response.status}`
+      );
+      error.status = response.status;
+      error.response = errorData || undefined;
+      throw error;
+    }
+
+    try {
+      const data = await response.json();
+      return data;
+    } catch {
+      throw new Error("Failed to parse response JSON");
+    }
+  }
+
+  /**
+   * Log request ใน development mode
+   */
+  private logRequest(method: string, url: string, data?: unknown): void {
+    if (process.env.NEXT_PUBLIC_DEV_MODE === "true") {
+      let detail: { method: string; url: string; data?: unknown | null } = {
+        method: method.toUpperCase(),
+        url,
+      };
+
+      if (data) detail.data = data;
+
+      console.log("🚀 API Request:", detail);
+    }
+  }
+
+  /**
+   * GET Request
+   */
+  async get<T = unknown>(
+    url: string,
+    config?: FetchConfig
+  ): Promise<ApiResponse<T>> {
+    const fullUrl = `${this.baseURL}${url}`;
+    this.logRequest("GET", fullUrl);
+
+    const response = await this.fetchWithTimeout(fullUrl, {
+      ...this.getDefaultConfig(),
+      ...config,
+      method: "GET",
+    });
+
+    return this.handleResponse<T>(response);
+  }
+
+  /**
+   * POST Request
+   */
+  async post<T = unknown>(
+    url: string,
+    data?: unknown,
+    config?: FetchConfig
+  ): Promise<ApiResponse<T>> {
+    const fullUrl = `${this.baseURL}${url}`;
+    this.logRequest("POST", fullUrl, data);
+
+    const response = await this.fetchWithTimeout(fullUrl, {
+      ...this.getDefaultConfig(),
+      ...config,
+      method: "POST",
+      body: data ? JSON.stringify(data) : undefined,
+    });
+
+    return this.handleResponse<T>(response);
+  }
+
+  /**
+   * PUT Request
+   */
+  async put<T = unknown>(
+    url: string,
+    data?: unknown,
+    config?: FetchConfig
+  ): Promise<ApiResponse<T>> {
+    const fullUrl = `${this.baseURL}${url}`;
+    this.logRequest("PUT", fullUrl, data);
+
+    const response = await this.fetchWithTimeout(fullUrl, {
+      ...this.getDefaultConfig(),
+      ...config,
+      method: "PUT",
+      body: data ? JSON.stringify(data) : undefined,
+    });
+
+    return this.handleResponse<T>(response);
+  }
+
+  /**
+   * DELETE Request
+   */
+  async delete<T = unknown>(
+    url: string,
+    config?: FetchConfig
+  ): Promise<ApiResponse<T>> {
+    const fullUrl = `${this.baseURL}${url}`;
+    this.logRequest("DELETE", fullUrl);
+
+    const response = await this.fetchWithTimeout(fullUrl, {
+      ...this.getDefaultConfig(),
+      ...config,
+      method: "DELETE",
+    });
+
+    return this.handleResponse<T>(response);
+  }
+
+  /**
+   * Download File (สำหรับ CSV, Excel, PDF etc.)
+   * @param url - endpoint URL
+   * @param filename - ชื่อไฟล์ที่ต้องการดาวน์โหลด
+   * @param config - fetch configuration
+   */
+  async download(
+    url: string,
+    filename: string,
+    config?: FetchConfig
+  ): Promise<void> {
+    const fullUrl = `${this.baseURL}${url}`;
+    this.logRequest("GET", fullUrl);
+
+    // สร้าง headers พิเศษสำหรับการดาวน์โหลดไฟล์
+    const downloadHeaders = {
+      ...this.getDefaultHeaders(),
+      Accept: "text/csv,application/csv,application/octet-stream,*/*",
+    };
+
+    const response = await this.fetchWithTimeout(fullUrl, {
+      ...this.getDefaultConfig(),
+      ...config,
+      method: "GET",
+      headers: {
+        ...downloadHeaders,
+        ...config?.headers,
+      },
+    });
+
+    if (!response.ok) {
+      let errorMessage = `HTTP Error ${response.status}: ${response.statusText}`;
+
+      try {
+        const errorData = await response.json();
+        errorMessage = errorData.message || errorMessage;
+      } catch {
+        // ถ้า parse JSON ไม่ได้ ใช้ error message เดิม
+      }
+
+      throw new Error(errorMessage);
+    }
+
+    // แปลงเป็น blob
+    const blob = await response.blob();
+
+    // สร้าง blob URL และดาวน์โหลดไฟล์
+    const blobUrl = window.URL.createObjectURL(blob);
+
+    // สร้าง element <a> เพื่อดาวน์โหลด
+    const link = document.createElement("a");
+    link.href = blobUrl;
+    link.download = filename;
+    document.body.appendChild(link);
+    link.click();
+
+    // ทำความสะอาด
+    document.body.removeChild(link);
+    window.URL.revokeObjectURL(blobUrl);
+  }
+}
+
+/**
+ * Singleton Instance - ใช้ instance เดียวทั่วทั้งแอพ
+ */
+const apiClient = new ApiClient();
+
+export default apiClient;
+
+/**
+ * Get full API URL with /api/v1 path
+ */
+export const getApiUrl = (path: string = ""): string => {
+  const baseUrl = getBaseUrl();
+  const apiPath = `/api/v1${path}`;
+  return `${baseUrl}${apiPath}`;
+};
