+++ lib/performance.ts
@@ -28,31 +28,6 @@ export function useDebounce<T>(value: T, delay: number): T {
   return debouncedValue;
 }
 
-// 🎯 Throttle Hook
-/**
- * Throttle hook that limits the rate at which a function can fire
- * @param callback - Function to throttle
- * @param delay - Delay in milliseconds
- * @returns Throttled function
- */
-export function useThrottle<T extends (...args: unknown[]) => unknown>(
-  callback: T,
-  delay: number
-): T {
-  const lastCall = useRef<number>(0);
-
-  return useCallback(
-    (...args: Parameters<T>) => {
-      const now = Date.now();
-      if (now - lastCall.current >= delay) {
-        lastCall.current = now;
-        return callback(...args);
-      }
-    },
-    [callback, delay]
-  ) as T;
-}
-
 // 🔄 Previous Value Hook
 /**
  * Hook to get the previous value of a state or prop
@@ -103,165 +78,6 @@ export function useStableCallback<T extends (...args: unknown[]) => unknown>(
 ): T {
   // eslint-disable-next-line react-hooks/exhaustive-deps
   return useCallback(callback, deps);
-}
-
-// 🏗️ Component Lazy Loading Utilities
-/**
- * Creates a lazy-loaded component with error boundary support
- * @param importFunc - Dynamic import function
- * @param fallback - Loading fallback component
- * @returns Lazy component
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function createLazyComponent<T extends React.ComponentType<any>>(
-  importFunc: () => Promise<{ default: T }>,
-  fallback?: React.ComponentType
-) {
-  const LazyComponent = React.lazy(importFunc);
-  const FallbackComponent =
-    fallback ||
-    (() => React.createElement("div", { className: "loading-skeleton h-32" }));
-
-  // Component wrapper
-  // eslint-disable-next-line react/display-name
-  const LazyWrapper = (props: React.ComponentProps<T>) =>
-    React.createElement(
-      React.Suspense,
-      { fallback: React.createElement(FallbackComponent) },
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      React.createElement(LazyComponent as any, props)
-    );
-
-  LazyWrapper.displayName = `LazyComponent`;
-
-  return LazyWrapper;
-}
-
-// 📱 Intersection Observer Hook for Lazy Loading
-/**
- * Hook for detecting when an element enters the viewport
- * @param options - Intersection Observer options
- * @returns [ref, isIntersecting]
- */
-export function useIntersectionObserver(
-  options: IntersectionObserverInit = {}
-): [React.RefObject<HTMLElement | null>, boolean] {
-  const [isIntersecting, setIsIntersecting] = useState(false);
-  const ref = useRef<HTMLElement | null>(null);
-
-  useEffect(() => {
-    const element = ref.current;
-    if (!element) return;
-
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        setIsIntersecting(entry.isIntersecting);
-      },
-      {
-        threshold: 0.1,
-        rootMargin: "50px",
-        ...options,
-      }
-    );
-
-    observer.observe(element);
-
-    return () => {
-      observer.unobserve(element);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return [ref, isIntersecting];
-}
-
-// 🎯 Lazy Loading Component
-/**
- * Component that only renders children when visible in viewport
- */
-interface LazyLoadProps {
-  children: React.ReactNode;
-  fallback?: React.ReactNode;
-  className?: string;
-  threshold?: number;
-  rootMargin?: string;
-}
-
-export function LazyLoad({
-  children,
-  fallback,
-  className,
-  threshold = 0.1,
-  rootMargin = "50px",
-}: LazyLoadProps): React.ReactElement {
-  const [ref, isIntersecting] = useIntersectionObserver({
-    threshold,
-    rootMargin,
-  });
-
-  const defaultFallback = React.createElement("div", {
-    className: "loading-skeleton h-32",
-  });
-
-  return React.createElement(
-    "div",
-    { ref, className },
-    isIntersecting ? children : fallback || defaultFallback
-  );
-}
-
-// 🔄 Async State Hook with Caching
-/**
- * Hook for managing async operations with caching
- * @param asyncFunction - Async function to call
- * @param deps - Dependencies for the async function
- * @param cacheKey - Optional cache key
- * @returns [data, loading, error, refetch]
- */
-export function useAsyncWithCache<T>(
-  asyncFunction: () => Promise<T>,
-  deps: React.DependencyList,
-  cacheKey?: string
-): [T | null, boolean, Error | null, () => void] {
-  const [data, setData] = useState<T | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
-
-  const cache = useRef<Map<string, T>>(new Map());
-
-  const execute = useCallback(async () => {
-    try {
-      setLoading(true);
-      setError(null);
-
-      // Check cache first
-      if (cacheKey && cache.current.has(cacheKey)) {
-        setData(cache.current.get(cacheKey)!);
-        setLoading(false);
-        return;
-      }
-
-      const result = await asyncFunction();
-
-      // Store in cache
-      if (cacheKey) {
-        cache.current.set(cacheKey, result);
-      }
-
-      setData(result);
-    } catch (err) {
-      setError(err instanceof Error ? err : new Error("Unknown error"));
-    } finally {
-      setLoading(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [...deps, cacheKey]);
-
-  useEffect(() => {
-    execute();
-  }, [execute]);
-
-  return [data, loading, error, execute];
 }
 
 // 🎮 Performance Monitoring
